Provide "python-docx" from pypi
This commit replaces legacy dependency repo for python-docx by pypi
releases.

It drops support for python 3.3, as no *.whl file exists for v0.8.11,
which is the most recent version that supports 3.3.

Can/Will add back the legacy dependency repo for python 3.3 in a
follow-up commit, if needed (will discuss in PR).